Add LLVM 21 apt repo for Ubuntu, fix MSVC initializer list error

+++ simulator/core/sim_value_parser.cpp
@@ -14,7 +14,7 @@ std::string trim(std::string_view sv) {
     auto end = sv.end();
     --end;
     while (end != start && std::isspace(static_cast<unsigned char>(*end))) --end;
-    return std::string(start, end + 1 - start);
+    return std::string(start, static_cast<size_t>(end + 1 - start));
 }
 
 std::string toLower(std::string_view sv) {
